refactor(utils): split diff formatting functions into multiple files
